Time: 28 ms (21.89%), Space: 17.2 MB (96.03%) - LeetHub

+++ 0437-path-sum-iii/0437-path-sum-iii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
 
-    int dfs(TreeNode* root, long long targetSum) {
+    int dfs(TreeNode* root, long targetSum) {
         if (!root) return 0;
         return (root->val == targetSum) + dfs(root->left, targetSum - root->val) + dfs(root->right, targetSum - root->val);
     }
